Fix/endpoint reviews

+++ routes/reviews/controllers.js
@@ -17,11 +17,20 @@ export const getAllReviews = async (_, res) => {
   }
 }
 
-// Get all Reviews
+// Get One Review
 export const getOneReview = async (req, res) => {
   try {
-    const id = parseInt(req.params.id)
+    const id = Number.parseInt(req.params.id, 10)
+
+    if (Number.isNaN(id)) {
+      return res.status(400).json({ message: 'Invalid booking id' })
+    }
+
     const result = await Review.findByPk(id)
+    if (!result) {
+      return res.status(404).json({ message: 'Review not found' })
+    }
+
     return res.status(200).json(result)
   } catch (error) {
     if (error instanceof Error) {
@@ -35,11 +44,18 @@ export const getOneReview = async (req, res) => {
 export const createReview = async (req, res) => {
   try {
     const body = req.body
-    const review = await createReviewSchema.validateAsync(body)
+
+    const review = await createReviewSchema.validateAsync(body, {
+      abortEarly: false,
+    })
     const result = await Review.create(review)
-    return res.status(200).json(result)
+
+    return res.status(201).json(result)
   } catch (error) {
     if (error instanceof Error) {
+      if ('details' in error) {
+        return res.status(422).json(error.details)
+      }
       return res.status(400).json({ message: error.message })
     }
     return res.status(500).json({ message: 'An error has occured' })
@@ -49,24 +65,29 @@ export const createReview = async (req, res) => {
 // Update Review
 export const updateReview = async (req, res) => {
   try {
-    const id = parseInt(req.params.id)
-    const body = req.body
-    const reviewUpdate = await updateReviewSchema.validateAsync(body)
+    const id = Number.parseInt(req.params.id, 10)
+    if (Number.isNaN(id)) {
+      return res.status(400).json({ message: 'Invalid booking id' })
+    }
 
-    const result = await Review.findOne({ where: { review_id: id } })
-    if (!result) {
+    const body = req.body
+    const validatedUpdate = await updateReviewSchema.validateAsync(body, {
+      abortEarly: false,
+    })
+
+    const review = await Review.findByPk(id)
+    if (!review) {
       return res.status(404).json({ message: 'Not found' })
     }
 
-    await Review.update(reviewUpdate, {
-      where: {
-        review_id: id,
-      },
-    })
+    await review.update(validatedUpdate)
 
-    return res.status(200).json({ message: 'Review updated' })
+    return res.status(200).json({ message: 'Review updated', review })
   } catch (error) {
     if (error instanceof Error) {
+      if ('details' in error) {
+        return res.status(422).json(error.details)
+      }
       return res.status(400).json({ message: error.message })
     }
     return res.status(500).json({ message: 'An error has occured' })
@@ -76,12 +97,18 @@ export const updateReview = async (req, res) => {
 // Delete Review
 export const deleteReview = async (req, res) => {
   try {
-    const id = parseInt(req.params.id)
-    const result = await Review.destroy({ where: { review_id: id } })
-    if (!result) {
-      return res.status(404).json({ message: 'Not found' })
+    const id = Number.parseInt(req.params.id, 10)
+
+    if (Number.isNaN(id)) {
+      return res.status(400).json({ message: 'Invalid booking id' })
     }
-    return res.status(200).json(result)
+
+    const deletedCount = await Review.destroy({ where: { review_id: id } })
+    if (!deletedCount) {
+      return res.status(404).json({ message: 'Review Not found' })
+    }
+
+    return res.status(200).json({ message: 'Review deleted', deletedCount })
   } catch (error) {
     if (error instanceof Error) {
       return res.status(400).json({ message: error.message })
